use `? extends Entity` bounded parameter type instead of just `Entity`
I faced a trouble using the service from Java code. It's not convenient and even it may be
impossible to pass `Class<ExtEntity>` to method declared to take `Class<Entity>`. But this
works in Groovy in your examples. So I decided to move to bounded generics and I found
them works well here. Since I have almost no experience in Groovy I migrated the service to
Java, but actually no one minds how it's implemented, the interface is what makes sense.